indent send float array insteady of byte array for CMV1 type

+++ host/pixymon/renderer.h
@@ -41,7 +41,7 @@ private:
     int renderCCQ1(uint8_t renderFlags, uint16_t width, uint16_t height, uint32_t numVals, uint32_t *qVals);
     int renderBA81(uint8_t renderFlags, uint16_t width, uint16_t height, uint32_t frameLen, uint8_t *frame);
     int renderCCB1(uint8_t renderFlags, uint16_t width, uint16_t height, uint32_t numBlobs, uint16_t *blobs);
-    int renderCMV1(uint8_t renderFlags, uint32_t cmodelsLen, uint8_t *cmodels, uint16_t width, uint16_t height, uint32_t frameLen, uint8_t *frame);
+    int renderCMV1(uint8_t renderFlags, uint32_t cmodelsLen, float *cmodels, uint16_t width, uint16_t height, uint32_t frameLen, uint8_t *frame);
     void emitImage(const QImage &image);
 
     int renderBA81Filter(uint16_t width, uint16_t height, uint32_t frameLen, uint8_t *frame);
